requirements.txt
fastmcp>=0.1.12
requests>=2.28.1
python-dotenv>=1.0.1
